refactoring and fix file names sort
